Change /bin to /util. Moved custom code for second settings to /src. Closes #7
