+++ src/chartfoundry/util/Renderer.js
