+++ src/chartfoundry/util/reactTemplate.js
