+++ src/chartfoundry/util/string-accessor.js
