+++ src/util/object-assign.js
